+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Changed equation used to calculate index of refraction for use switched from
+  one provided in Stanley Gedzelman's paper Simulatin Rainbows and Halos in
+  Color to a polynomial fitted to data provided by Warren and Brandt in their
+  paper Optical Constants of Ice from The Ultraviolet to The Microwave, which
+  is what HaloPoint 2.0 used to use
+
 ## 4.0.1 - 2021-08-26
 
 ### Fixed