+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.1.0 - 2022-07-23
+
+### Changed
+
+- Changed equation used to calculate index of refraction for use switched from
+  one provided in Stanley Gedzelman's paper Simulatin Rainbows and Halos in
+  Color to a polynomial fitted to data provided by Warren and Brandt in their
+  paper Optical Constants of Ice from The Ultraviolet to The Microwave, which
+  is what HaloPoint 2.0 used to use
+- HaloRay now remembers which folder you opened last when loading or saving
+  files like simulation parameter files or images, and opens the file dialog
+  picker in that same folder the next time
+
 ## 4.0.1 - 2021-08-26
 
 ### Fixed